Added clearer explanation of mission and tensions
I put the mission of Hacker Society at the top. I changed the
explanation of why we don't allow some behaviour. I clarified the
section on the tension between restrictions on behaviour and expression
and the goals of an open society. Explicitly noting that: unpopular
ideas are allowed, politically incorrect speach is allowed, distastful
speach is allowed. The line is where that speach becomes an attack on a
person in the community and is about one or more of their intrinstic
characteristics.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>